requirements.txt
langchain-core==0.1.17
langchain-openai==0.0.5
openai==1.14.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1